Land some CMake revisions for wasm integration
Cherry-picking some CMake-related stuff from #5097 to make it easier to review:

- Add an optional COMMAND argument to add_halide_test(), this allows us to customize the command for execution (e.g. to run generated .wasm with a shell tool)

- Add some missing DEPENDS in a few places

- Add an optional EXTRA_LIBS argument to halide_define_aot_test(); this allows us to pass extra dependencies rather than requiring separate calls to target_link_libraries().

That last one is a little odd, so let me expand: the intent here is that (when the wasm changes land), some of the tests that aren't usable under wasm (e.g. matlab), and halide_define_aot_test() will handle these by just skipping those targets entirely. This means that we can effectively centralize the blocklist in one place, and then the callers of halide_define_aot_test() can just do something like

        halide_define_aot_test(matlab)
        if (TARGET generator_aot_matlab)
          set_target_properties(generator_aot_matlab PROPERTIES ENABLE_EXPORTS True)
        endif ()

... i.e., just assume that if the target may not be defined if it's blacklisted for some reason.

(Open for better suggestions on this last one, but I felt it was better than spraying lots of checks for "if wasm blah blah" thru this entire file)