Add missing overload for boundary conditions on a buffer
The following syntax did not work due to a lack of appropriate overload:

repeat_edge(buffer, {{0, W}, {0, H}})

The version that worked was:

repeat_edge(buffer, 0, W, 0, H)

but that syntax has fallen out of favor

This PR adds an overload to make the former syntax work, and explicitly
deprecates the latter.

Also removed inline qualifiers on some templated functions, because
they're already inline, and changed an arg to be by const ref instead of
pointlessly by value.