Add a Target to the args of infer_input_bounds()
Currently, if the code isn't already jitted (or is jitted with NoBoundsQuery enabled), infer_input_bounds() falls back to whatever is in HL_JIT_TARGET. This is suboptimal, because it hides the fact that infer_input_bounds() relies on the value of that env var. It can also cause unintentional incorrect re-jits; e.g. if simd_op_check is run for HL_TARGET=wasm-32-wasmrt (and HL_JIT_TARGET not set at all), the infer_input_bounds() call can end up re-jitting for native code instead of WebAssembly.