Remove unnecessary call to halide.hexagon.pack.vh in CG_HVX
In the degenerate case of shuffle_vector() calling vlut() to shuffle a vector that is wider than 256 elements, the code was incorrectly using halide.hexagon.pack.vh on a vector-of-bool; this used to be necessary, but hasn't been for a while, so clearly this code path wasn't being exercised.

Remove the halide.hexagon.pack.vh and added a test case to exercise that path.

Also, drive-by removal of #include "EliminateBoolVectors.h" from CGHVX since it is no longer used there.