Add infer_input_bounds(vector<int>)
Add a variant of infer_input_bounds() that takes an explicit vector of int, rather than the up-to-4-int version that is a holdover from the buffer_t days; deprecate the old version; convert all existing code to use the new one.

Note that I'm using a new overload (with an initializer-list) as a way to subvert the mis-binding of `{}` and `{1}` to the deprecated function; this adds a trivial amount of overhead but (I think) allows us to ensure that converted code probably avoids the deprecated method.