Clean up is-jit-compiled checks in Pipeline (#5172)
* Clean up is-jit-compiled checks in Pipeline

Because WebAssembly is a special beast, the way it is 'jitted' is special, and the checks to avoid redundant jitting needed extra logic in compile_jit(). Unfortunately there was another place in Pipeline that also needed this special casing. This PR adds a `get_compiled_jit_target()` bottleneck to consolidate this.

* defined() -> has_unknowns()