Sketch of statically-typed Exprs
This is a a rough sketch of an idea I have been pondering: could we usefully add more (optional) static typing to the Halide C++ IR?

This PR just adds a wrapper around `Expr` to allow for static typing (a la `Buffer<>`). The idea here is that an `ExprT<type>` is an Expr that can only a value of the given type (or be undefined).

The main goal here is to allow for easier readability in complex pipelines (eg if an Expr is expected to be of type uint8, this makes it more plain when reading the source code, and failures would happen at compile time, not runtime).

I'm not entirely sure if I like this or not, so I didn't expend a lot of energy trying to complete it -- there may be holes in the implementation.

In the name of science, I modified `cast<T>()` to return `ExprT<T>`, and `Var` to be convertible to `ExprT<int>` (rather than plain `Expr`).

If this seems remotely desirable, perhaps it could lead to similar wrappers for Func, ImageParam, etc.

Thoughts welcome.