Avoid possibly-uninitialized use of RVar::_index
One of the armbots warned that this field could be used uninitialized; I can't replicate anywhere else, but indeed, the string-only ctor of RVar left this uninitialized. Defaulted it to -1 and added an explicit check in _var(). (Yes, the call to `at()` will fail when out of range, but explicit checking is better IMHO.)