Handle more types of nested vectorized += without scalarizing
With this change, in a nesting of vectorized vars in an
associative/commutative reduction (e.g. +=), we can now have a reduction
var outermost and a reduction var innermost and get good codegen. This
is still not fully general - there can only be one pure var in the
vectorized stack for it to work. In general is_interleaved_ramp should
be an is_tensor_contraction pass that knows how to do clever codegen for
those.

For the following schedule:

```
              prod.compute_at(result, x)
                    .vectorize(x)
                    .update()
                    .split(r, ro, ri, 8)
                    .split(ri, rio, rii, 2)
                    .reorder(rii, x, rio, ro)
                    .vectorize(x)
                    .atomic()
                    .vectorize(rio)
                    .vectorize(rii);
```

We get the following IR:

```
let t2262 = (int32x32)vector_reduce(Add,
(int32x64(shuffle((int16x15)p10_im_global_wrapper$0[ramp(0, 1, 15)], 0,
1, 1, 2, 2, 3, 3, 4, 4, 5, 5, 6, 6, 7, 7, 8, 2, 3, 3, 4, 4, 5, 5, 6, 6,
7, 7, 8, 8, 9, 9, 10, 4, 5, 5, 6, 6, 7, 7, 8, 8, 9, 9, 10, 10, 11, 11,
12, 6, 7, 7, 8, 8, 9, 9, 10, 10, 11, 11, 12, 12, 13, 13,
14))*int32x64(shuffle((int16x8)p11_im_global_wrapper$0[ramp(0, 1, 8)],
0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 0, 1, 2, 3, 2, 3, 2, 3, 2, 3,
2, 3, 2, 3, 2, 3, 2, 3, 4, 5, 4, 5, 4, 5, 4, 5, 4, 5, 4, 5, 4, 5, 4, 5,
6, 7, 6, 7, 6, 7, 6, 7, 6, 7, 6, 7, 6, 7, 6, 7))))

f58[ramp(0, 1, 8)] = slice_vectors(t2262, 8, 1, 8) +
(slice_vectors(t2262, 0, 1, 8) + (slice_vectors(t2262, 16, 1, 8) +
(slice_vectors(t2262, 24, 1, 8) + f58[ramp(0, 1, 8)])))

```

The vector_reduce is from rii, and the sum of slice_vectors is from rio.

The generated asm (for avx512) is:

```
	leal	(%rbx,%rcx), %edx
	movslq	%edx, %rdx
	subq	%r14, %rdx
	vmovdqu	(%r15,%rcx,2), %xmm6
	vbroadcasti64x4	(%r12,%rdx,2), %zmm7 # zmm7 =
mem[0,1,2,3,0,1,2,3]
	vpermw	%zmm7, %zmm0, %zmm8
	vpermw	%zmm7, %zmm1, %zmm7
	vpermd	%zmm6, %zmm2, %zmm6
	vpermd	%zmm6, %zmm3, %zmm9
	vpmaddwd	%zmm9, %zmm8, %zmm8
	vpermd	%zmm6, %zmm4, %zmm6
	vpmaddwd	%zmm6, %zmm7, %zmm6
	vextracti64x4	$1, %zmm6, %ymm7
	vextracti64x4	$1, %zmm8, %ymm9
	vpaddd	%ymm5, %ymm8, %ymm5
	vpaddd	%ymm6, %ymm5, %ymm5
	vpaddd	%ymm5, %ymm9, %ymm5
	vpaddd	%ymm7, %ymm5, %ymm5
	addq	$8, %rcx
	cmpq	$128, %rcx
```

The pmaddwds are from rii, and the vpaddds are from rio.

This is 3.5x faster than the best schedule that only vectorizes the pure
var, and about 2x faster than the best schedule that only vectorizes an
unsplit reduction variable.