Add logical_not function for Python
This change introduces `logical_not` as a free function and member function
that calls `operator!`.

The reason why a new function is added is because there is no `operator!`
in Python and the `not` keyword cannot be overloaded. Hence, there was
currently no way to call the C++ `operator!` in Python.