Push lets near their uses (whenever possible) while CSEing +=
The CSE pass tries to do CSE jointly on the index and value of store nodes.
This is to stop:
f[x] = f[x] + y
from turning into
f[x] = f[z] + y
This is because the two 'x' indices are not CSE'd together.

However, one problem with jointly CSEing the store index and vaues
is that after CSEing, the pass puts the lets
(the values that were CSE'd) before the new store. That is we get,

let(t0...)
let(t1...)
f[t0] = f[t0] + function_of(t1)

Suppose however, there was nothing to CSE between the store index and value, that is
the index was unchanged after CSE. In that case, moving lets before the store puts the
lets too far away from their uses. This is ok except, there are passes like LoopCarry
that are beneficial when they are able to see a long continuous block of stores (eg. when
unrolling). But now, they'll see a long sequence of LetStmts.
Instead, if the index was unchanged after CSE, we should build up the stores as

f[x] = let(t0.. in (value));

This way LoopCarry is given a chance to see a series of stores.
Handling this in CSE, means the LoopCarry pass need not be complicated.

Change-Id: Iae19e1f69a6b38f3224a64b0c4781533e3862970