Be smarter about pushing lets near their uses when CSEing +=
In a previous patch, in the code that handles CSEing += store operations,
we weren't general enough; we only handled the specific case when the
store index remained unchanged when CSE'd together with the store value.

This patch is more general. Even if the store index changes, only the
use-def chains created from the values used by the new store index that
end in defs created by CSE are retained as Let stmts around the new store.
Others go as Let expressions around the store value.