Modify memoization code to allow using min/extent/stride of a buffer as part of memoization without wrapping in memoize_tag. This seems reasonable and failure to support this causes tricky to diagnose errors if one uses the extent of an input in an RDom that is then used in a memoized Func.
The code pattern here is a bit heuristic in that I can't think of a
case where a Var has a buffer but the rference isn't to a field of the
buffer. This is something to think about in review. I could also
pattern match the variable name, but that seems worse.