Add initial support for building a TFLite Delegate
This builds a TFLite delegate as a shared library, and modifies compare_vs_tflite (and run_compare_on_device) to run *three* ways by default:
- vs "plain" TFLite
- vs TFLite when using our Delegate
- calling our Halide code 'directly'

This is still very rough and has oodles of rough edges and TODOs for things that need improvement, but it does run mobilenetv2 correctly (on desktop and on android). Pushing this out as a PR to preserve work over vacation (and for Dillon to take a look if he likes) but no pressure to land this just yet -- it's still a WIP so it can wait to land if we like (or land now with the understanding that it's still very rough).