Remove OpenGL support (part 1)
Fixes #5475

This removes the OpenGL backend (but *not* the OpenGLCompute backend) from public use:

- Remove Target::OpenGL
- remove DeviceAPI::GLSL
- remove Func::glsl() and Func::shader()
- remove all OpenGL-specific apps and tests
- remove HalideRuntimeOpenGL.h
- remove some internal code that is OpenGL-only

Note that there is still internal code that needs trimming; since the OpenGLCompute backend uses some of the same code, and some of the same build deps, and some of the same runtime shared-library loading, I tried to err on the side of leaving code/buildrules/etc in place for now, with the plan to clean that up in subsequent PRs.

Note also that feature Target::EGL is still present, as I believe it is still useful in conjunction with OpenGLCompute.