Make write_debug_image MSAN-safe
If you build an msan-enabled test, you can end up with stack space allocated by the caller marked as "invalid" (e.g. if they allocate a decent buffer on the stack and only use a few bytes of it). If this test calls halide_debug_to_file(), the stack space used internally by that function can overlay the stack space from the caller (which is fine), but since our runtime code is never compiled with msan enabled, the writes to this stack space in our code never set the shadow bits... which is fine until a system call (like fwrite) is called, and then msan blows up.

The "solution" for this is to just mark the things handed to fwrite (etc) as explicitly initialized; note that *everything* calling this is going thru a stack-based buffer of some sort (we never use heap memory directly).