Refactor code for dealing with default values of scalar Params
The "default" value for a scalar param is rarely used -- it's currently only possible to specify foran Input in a Generator, and that value only shows up in the generated metadata for AOT compilation.

This PR refactors this so that instead of being maintained solely as a hack in Generator data structures, it's moved into Parameter as its own field.

This seems like a lot of work some something of marginal use, but I'm reluctant to suggest removing it entirely (it's possible it could break someone's code), and refactoring it in this way will make some subsequent Generator refactoring easier to understand and review.