Do something like sqrt-free Cholesky for BGU (#5281)
* Do something like sqrt-free Cholesky for BGU

This produces fewer update definitions and less total math in the solve
step, saving some code size. No significant impact on performance for
this app, because the solve step is done at low res, but it's
theoretically more satisfying, and it's nice to have the symmetric
version of the matrix solve available as a reference.