Constrain calling convention for PyStub (#5761)
There are two calling conventions that are sensible when calling a Generator from Python:
- Inputs are specified via keyword arguments (verbose but more readable)
- Inputs are specified via positional arguments (terse but more like C++)

However, the current PyStub implementation also allows for mixing positional and keyword forms of input (as long as no input is specified multiple times). This change removes that possibility (requiring either all-keyword or all-positional for inputs). I'm not sure why I thought this was a good option originally, but upon reflection (and examination of existing code), calls written in this way tend to be a confusing mess that require too much care when reading, and I think we'd be better off just forbidding this entirely.

Note that this technically is a breaking API change; any user code that used this technique previously would now throw an exception and need a (trivial) update. That said, this is (AFAIK) a very rarely used API, and I rather suspect that existing code that relies on being able to mix positional and keyword inputs in this way is likely to be doing so inadvertently rather than deliberately.