Add a way to set a GeneratorInput's type in code (#5868)
* Add a way to set a GeneratorInput's type in code

Currently, if you want to vary the type of a Generator's inputs or outputs, you have to specify the types in the makefile. This can be awkward for things with complex logic. This PR proposes adding a way to do this: a new `set_type()` method which can only be called from the rarely-used Generator::configure() method. It only allows setting the type for an input or output that has no type specified.

I'm not 100% sure if this is a good idea, but for certain rare corner cases, it may be quite handy.

(Note that extending this to allow specifying dimensions and/or array size in the same way might be handy, but is omitted from this PR.)

* Update Generator.h

* Also add set_dimensions, set_array_size