Fix an issue in Halide's float16 compilation support. Add tests. (#5872)
In EmulateFloat16Math.cpp, conversion from 32-bit float to 16-bit float could produce a NaN value when and infinity is correct. This is because for numbers larger than the exact infinity value, the mantissa could be non zero.

Add tests to cover this case, and float16 infinities in general.

Couple small style/comment cleanups.