Initial work to add Tensor Cores support in Halide
Tensor Cores are programmable cores that perform warp-level
matrix-multiply and accumulate operations. It can greatly improve
performance of matrix-multiply expression written in Halide when
the hardware is available.

This works by pattern-matching an expression of the form

RDom k(0, matrix_size);
C(x, y) += f32(A(k, y) * B(x, k));

and generating code that will realize the function using the GPU.

The current support is quite limited. The only data types currently
support are float16_t for A and B and float for C. The dimentions of
the input matrices must be a multiple of 16 since the only shape
currently supported is m16n16k16.

The following table shows the performance comparison of the modified
matrix_multiplication performance test for different configurations

| Configuration (Matrix Size: 992) | Time (ms) | GFlops  |
|----------------------------------|-----------|---------|
| CPU (Vectorized)                 | 148.346   | 13.16   |
| CUDA                             | 4.345     | 449.23  |
| CUDA (Tensor Cores)              | 1.88      | 1038.48 |

CPU: Intel Core i7-8700
GPU: NVIDIA RTX2060