Avoid pathological cases in halide_benchmark()
In the variant that tries to compute a good samples/iters value based on min_time, there's a pathological case if the environment's timer is relatively coarse, and the op being profiled is relatively fast; in this case, you can end up with timings that are very close to zero (or *literally* zero), and our attempt to calculate the number of iterations can explode into the billions, making the benchmark appear to hang (as it may take an absurd length of time to run). To fix this, add a maximum value for iters_per_sample, and smarten the calculation for when the measured time is tiny.