[hannk] Proposal for better error handling
Currently, hannk deals with most errors by crashing (either via HCHECK or assert).

This proposes a Status return code approach:
- In opt builds, is just an int-sized class, marked with [[nodiscard]] so that it can't be accidentally ignored

- in non-opt builds, we also gather the function/file/line from the point of instantiation, so that we can eventually report that to aid with debugging.

- Made some initial changes in file_util.h and model_runner to use it.

The astute observer will note that this is not unlike a very very cut-down version of absl::Status, which is not an accident... except that this is intended to stay extremely minimal. (In point of fact, I was going to propose just using an enum marked with [[nodiscard]], but being able to optionally preserve source location, etc for debugging purposes seems like a worthwhile bonus at what should be negligible code cost in opt builds.)

If we like this approach, I'll likely implement it in a few PRs, adding Status results to the API surfaces that need it, changing callers as needed.

(I debated over whether an equivalent of absl::StatusOr would be useful or overkill, and ended up leaving it out. Thoughts?)