Add IRMutator::mutate_exprs()
There's a common pattern in many IRMutators that is "mutate a vector<Expr> and optionally let me know if anything is different".

(Note, this uses C++17 structured-binding syntax, which we previously weren't using in Halide. Objections?)

This adds a shared utility method (well, two, thanks to VariadicVisitor) and plus in the usage in all the places that seemed obvious.

I doubt this moves the needle on speed in either direction, but makes for smaller code.