Refactor mutate-vector-of-Expr code
There's a common pattern in many IRMutators that is "mutate a vector<Expr> and optionally let me know if anything is different".

This just adds `mutate()` and `same_as()` for `vector<Expr>` (and refactors some obvious places to use them).

I doubt this moves the needle in terms of performance, but IMHO it's cleaner code.