Hoist unrolled prefetches to top of the block (#6230)
* Hoist unrolled prefetches to top of the block

When a loop with prefetch is unrolled, the prefetch instructions getting scattered through the loop can cause LLVM codegen issues in some cases (see https://bugs.llvm.org/show_bug.cgi?id=51172). As a partial mitigation for that issue, this PR adds a pass to hoist all prefetch instructions to the top of their loop. This is still a bit experimental; it definitely addresses the codegen issues we see, but makes the use of prefetch potentially less effective (since the hoisted prefetch may be too far from the eventual use to be effective).

* appease clang-tidy

* Avoid quadratic behavior

* Use template instead of std::function

* Require prefetch offset to be pure